test(config): move and expand tests for empty_to_none helper
- Move `_empty_to_none` from `config.py` to `utils/formatting.py` as `empty_to_none`
- Update `config.py` to use the new utility
- Move and expand unit tests to `tests/test_utils/test_formatting.py`
- Achieve 100% coverage for the utility function
